feat: add formatPostSummary function and refactor post summary handling
